move IS_GPU/IS_CPU out of omniscidb
